# docs/swagger.yaml
basePath: /api/v1
definitions:
  operate.ModelError:
    properties:
      error_message:
        type: string
    type: object
...
  request.CreateUser:
    properties:
      login:
        example: login
        type: string
      password:
        example: password
        type: string
      role:
        default: user
        enum:
        - user
        - admin
        example: user
        type: string
    type: object
  request.Login:
    properties:
      login:
        example: login
        type: string
      password:
        example: password
        type: string
    type: object
...
  request.UpdateRole:
    properties:
      role:
        enum:
        - user
        - admin
        example: user
        type: string
    type: object
...
  response.User:
    properties:
      id:
        example: 5
        format: uint64
        type: integer
      login:
        example: login
        type: string
      role:
        enum:
        - user
        - admin
        example: user
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: vet_v2002@mail.ru
    name: Artem Vetoshkin
  description: Rest API Для управления базо данных фильмов
  title: Фильмотека
  version: "1.0"
paths:
...
  /login:
    post:
      consumes:
      - application/json
      description: Авторизация пользователя в системе.
      parameters:
      - description: Логин и пароль пользователя
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.Login'
      produces:
      - application/json
      responses:
        "200":
          description: Пользователь успешно авторизован
          headers:
            Set-Cookie:
              description: Устанавливает сессию текущего пользователя
              type: string
        "400":
          description: В теле запроса ошибка
          schema:
            $ref: '#/definitions/operate.ModelError'
        "409":
          description: Неверный логин или пароль
          schema:
            $ref: '#/definitions/operate.ModelError'
        "418":
          description: Пользователь уже авторизован
          schema:
            $ref: '#/definitions/operate.ModelError'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/operate.ModelError'
      summary: Авторизация.
      tags:
      - user
  /logout:
    post:
      description: Позволяет выйти пользователю из системы.
      produces:
      - application/json
      responses:
        "200":
          description: Пользователь успешно вышел из системы
        "401":
          description: Пользователь не авторизован
          schema:
            $ref: '#/definitions/operate.ModelError'
      security:
      - sessionCookie: []
      summary: Выход из системы.
      tags:
      - user
  /user:
    post:
      consumes:
      - application/json
      description: Добавляет пользователя включая его логин, пароль и роль. По умолчанию
        роль 'user'.
      parameters:
      - description: Информация о добавляемом пользователе
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.CreateUser'
      produces:
      - application/json
      responses:
        "201":
          description: Пользователь успешно добавлен в базу
          schema:
            $ref: '#/definitions/response.User'
        "400":
          description: В теле запроса ошибка
          schema:
            $ref: '#/definitions/operate.ModelError'
        "401":
          description: Пользователь не авторизован
          schema:
            $ref: '#/definitions/operate.ModelError'
        "403":
          description: У пользователя нет прав на создание пользователя
          schema:
            $ref: '#/definitions/operate.ModelError'
        "409":
          description: Пользователь с таким же логином уже существует
          schema:
            $ref: '#/definitions/operate.ModelError'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/operate.ModelError'
      security:
      - sessionCookie: []
      summary: Добавление пользователя.
      tags:
      - user
  /user/{user_id}:
    delete:
      description: Удаляет пользователя по его id.
      parameters:
      - description: Уникальный идентификатор пользователя
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Пользователь успешно удалён
        "400":
          description: В теле запроса ошибка
          schema:
            $ref: '#/definitions/operate.ModelError'
        "401":
          description: Пользователь не авторизован
          schema:
            $ref: '#/definitions/operate.ModelError'
        "403":
          description: У пользователя нет прав на удаление пользователя
          schema:
            $ref: '#/definitions/operate.ModelError'
        "404":
          description: Пользователь с указанным id не найден
          schema:
            $ref: '#/definitions/operate.ModelError'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/operate.ModelError'
      security:
      - sessionCookie: []
      summary: Удаление пользователя.
      tags:
      - user
  /user/{user_id}/role:
    put:
      consumes:
      - application/json
      description: Обновляет пользовательскую роль.
      parameters:
      - description: Уникальный идентификатор пользователя
        in: path
        name: user_id
        required: true
        type: integer
      - description: Информация о добавляемом пользователе
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.UpdateRole'
      produces:
      - application/json
      responses:
        "200":
          description: Роль пользователя успешно обновлена
          schema:
            $ref: '#/definitions/response.User'
        "400":
          description: В теле запроса ошибка
          schema:
            $ref: '#/definitions/operate.ModelError'
        "401":
          description: Пользователь не авторизован
          schema:
            $ref: '#/definitions/operate.ModelError'
        "403":
          description: У пользователя нет прав на создание пользователя
          schema:
            $ref: '#/definitions/operate.ModelError'
        "404":
          description: Пользователь с указанным id не найден
          schema:
            $ref: '#/definitions/operate.ModelError'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/operate.ModelError'
      security:
      - sessionCookie: []
      summary: Обновление роли пользователя.
      tags:
      - user
  /user/list:
    get:
      description: Возвращает список пользователей системы.
      produces:
      - application/json
      responses:
        "200":
          description: Список пользователей успешно сформирован
          schema:
            items:
              $ref: '#/definitions/response.User'
            type: array
        "401":
          description: Пользователь не авторизован
          schema:
            $ref: '#/definitions/operate.ModelError'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/operate.ModelError'
      security:
      - sessionCookie: []
      summary: Получение списка пользователей.
      tags:
      - user
schemes:
- http
securityDefinitions:
  sessionCookie:
    description: Уникальный идентификационный номер сессии
    in: cookie
    name: session_id
    type: apiKey
swagger: "2.0"